Hangman_game.py
This Hangman Python Project is a basic word-guessing game where the player has to guess the letters of a randomly chosen word from a predefined list. The player is given 6 chances to guess the correct word. For every wrong guess, the number of remaining chances decreases. Correct guesses reveal the letter's position(s) in the word. The game ends when the player either correctly guesses the word or runs out of chances. This project is great for beginners to practice loops, conditionals, user input handling, and string manipulation in Python.